[MIG] stock_operating_unit: Migration to v11
