+++ stock_operating_unit/data/stock_data.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2016-2017 Eficent Business and IT Consulting Services S.L.
-     Copyright 2016-2017 Serpent Consulting Services Pvt. Ltd.
+<!-- Copyright 2019 Eficent Business and IT Consulting Services S.L.
+     Copyright 2019 Serpent Consulting Services Pvt. Ltd.
      License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl-3.0) -->
 <odoo noupdate="1">
