+++ stock_operating_unit/tests/__init__.py
@@ -1,7 +1,6 @@
-# -*- coding: utf-8 -*-
-# © 2016-2017 Eficent Business and IT Consulting Services S.L.
+# © 2019 Eficent Business and IT Consulting Services S.L.
 # - Jordi Ballester Alomar
-# © 2016-2017 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
+# © 2019 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 from . import test_stock_operating_unit
 from . import test_stock_picking
