+++ stock_operating_unit/tests/test_stock_operating_unit.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
-# © 2016-2017 Eficent Business and IT Consulting Services S.L.
-# © 2016-2017 Serpent Consulting Services Pvt. Ltd.
+# © 2019 Eficent Business and IT Consulting Services S.L.
+# © 2019 Serpent Consulting Services Pvt. Ltd.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
-from openerp.addons.stock.tests import common
+from odoo.addons.stock.tests import common
 
 
 class TestStockOperatingUnit(common.TestStockCommon):
